+++ HISTORY.md
@@ -3,3 +3,7 @@
 <https://www.conventionalcommits.org/en/v1.0.0/>
 
 ## [unreleased]
+
+## [0.1.0] - 2026-02-04 | Initial commit / mount theft protection
+
+- feat: mount theft protection